requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.9.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
cssselect==1.2.0
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
Flask==3.0.3
Flask-Caching==2.3.0
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.2.29
langchain-text-splitters==0.2.2
langchain-openai==0.1.21
langsmith==0.1.98
lxml==5.2.2
lxml_html_clean==0.2.0
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
openai==1.40.1
orjson==3.10.7
packaging==24.1
pandas==2.2.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
reportlab==4.2.2
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
striprtf==0.0.26
tenacity==8.5.0
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
